+++ src/main/java/org/prgrms/nabimarketbe/domain/oauth2/google/api/GoogleOAuth2Controller.java
@@ -1,15 +1,15 @@
-package org.prgrms.nabimarketbe.oauth2.google.api;
+package org.prgrms.nabimarketbe.domain.oauth2.google.api;
 
 import java.io.IOException;
 
 import javax.servlet.http.HttpServletResponse;
 
+import org.prgrms.nabimarketbe.domain.oauth2.google.service.GoogleOAuth2Service;
 import org.prgrms.nabimarketbe.domain.user.dto.request.SocialUserInfoDTO;
 import org.prgrms.nabimarketbe.domain.user.dto.response.UserLoginResponseDTO;
 import org.prgrms.nabimarketbe.domain.user.service.SignService;
 import org.prgrms.nabimarketbe.global.util.ResponseFactory;
 import org.prgrms.nabimarketbe.global.util.model.SingleResult;
-import org.prgrms.nabimarketbe.oauth2.google.service.GoogleOAuth2Service;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
