+++ src/main/java/org/prgrms/nabimarketbe/domain/oauth2/google/dto/GoogleUserInfoDTO.java
@@ -1,0 +1,19 @@
+package org.prgrms.nabimarketbe.domain.oauth2.google.dto;
+
+import org.prgrms.nabimarketbe.domain.user.entity.Role;
+import org.prgrms.nabimarketbe.domain.user.entity.User;
+
+public record GoogleUserInfoDTO(
+    String id
+) {
+    private static final String PROVIDER = "GOOGLE";
+
+    public User toEntity(String nickName) {
+        return User.builder()
+            .nickname(nickName)
+            .provider(PROVIDER)
+            .accountId(id)
+            .role(Role.USER.getKey())
+            .build();
+    }
+}
