--- src/main/java/org/prgrms/nabimarketbe/oauth2/google/domain/Role.java
@@ -1,13 +1,0 @@
-package org.prgrms.nabimarketbe.oauth2.google.domain;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Role {
-	GUEST("ROLE_GUEST"),
-	USER("ROLE_USER");
-
-	private final String key;
-}
