--- src/main/java/org/prgrms/nabimarketbe/oauth2/google/dto/GoogleOAuth2TokenDTO.java
@@ -1,20 +1,0 @@
-package org.prgrms.nabimarketbe.oauth2.google.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record GoogleOAuth2TokenDTO(
-	@JsonProperty("access_token")
-	String accessToken,
-
-	@JsonProperty("expires_in")
-	int expiresIn,
-
-	String scope,
-
-	@JsonProperty("token_type")
-	String tokenType,
-
-	@JsonProperty("id_token")
-	String idToken
-) {
-}
